Clean up makefile and top level index creation script a bit

+++ bin/create_indexes_from_ucsc.pl
@@ -41,14 +41,14 @@ $F2 =~ s/.txt$/_one-line-seqs_temp.fa/;
 $F3 = $infile;
 $F3 =~ s/.txt$/_one-line-seqs.fa/;
 
-run_subscript "modify_fasta_header_for_genome_seq_database.pl", "$infile > $F1";
-run_subscript "modify_fa_to_have_seq_on_one_line.pl",  "$F1 > $F2";
-run_subscript "sort_genome_fa_by_chr.pl", "$F2 >  $F3";
+#run_subscript "modify_fasta_header_for_genome_seq_database.pl", "$infile > $F1";
+#run_subscript "modify_fa_to_have_seq_on_one_line.pl",  "$F1 > $F2";
+#run_subscript "sort_genome_fa_by_chr.pl", "$F2 >  $F3";
 
-unless ($debug) {
-  unlink($F1);
-  unlink($F2);
-}
+#unless ($debug) {
+#  unlink($F1);
+#  unlink($F2);
+#}
 
 $NAME = $ARGV[1];
 
@@ -101,10 +101,12 @@ open(OUTFILE, ">$configfile");
 print OUTFILE $config;
 close(OUTFILE);
 
-unlink("gene_info_merged_unsorted.txt");
-unlink("gene_info_merged_unsorted_fixed.txt");
-unlink("gene_info_merged_sorted_fixed.txt");
-unlink("master_list_of_exons.txt");
+unless ($debug) {
+  unlink("gene_info_merged_unsorted.txt");
+  unlink("gene_info_merged_unsorted_fixed.txt");
+  unlink("gene_info_merged_sorted_fixed.txt");
+  unlink("master_list_of_exons.txt");
+}
 
 # run bowtie on genes index
 print STDERR "\nRunning bowtie on the gene index, please wait...\n\n";
